% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_descriptive_plots.R
\name{analysis_descriptive_plots}
\alias{analysis_descriptive_plots}
\title{analysis_descriptive_plots}
\usage{
analysis_descriptive_plots(DF_joined, DF_raw, DF_clean, save_plots = FALSE)
}
\arguments{
\item{DF_joined}{.}

\item{DF_raw}{.}

\item{DF_clean}{.}

\item{save_plots}{TRUE / FALSE}
}
\description{
Generate histograms for main variables
}
