% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_model_E1.R
\name{analysis_model_E1}
\alias{analysis_model_E1}
\title{analysis_model_E1}
\usage{
analysis_model_E1(DF_analysis)
}
\arguments{
\item{DF_analysis}{DF_analysis}
}
\value{
lm model
}
\description{
analysis_model_E1
}
