% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_model_E1_plot.R
\name{analysis_model_E1_plot}
\alias{analysis_model_E1_plot}
\title{analysis_model_E1_plot}
\usage{
analysis_model_E1_plot(model)
}
\arguments{
\item{model}{model object}
}
\value{
ggplot
}
\description{
analysis_model_E1_plot
}
