% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_model_E1_table.R
\name{analysis_model_E1_table}
\alias{analysis_model_E1_table}
\title{analysis_model_E1_table}
\usage{
analysis_model_E1_table(model)
}
\arguments{
\item{model}{model object}
}
\value{
gtsummary table
}
\description{
analysis_model_E1_table
}
