% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{auto_reliability}
\alias{auto_reliability}
\title{auto_reliability}
\usage{
auto_reliability(
  DF,
  short_name_scale = short_name_scale_str,
  items = NULL,
  min_rdrop = 0.2
)
}
\arguments{
\item{DF}{.}

\item{short_name_scale}{.}

\item{items}{.}

\item{min_rdrop}{.}
}
\description{
Checks Cronbach's alpha and select variables based on a min r.drop criteria
}
