% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{check_NAs}
\alias{check_NAs}
\title{check_NAs}
\usage{
check_NAs(DF_joined)
}
\arguments{
\item{DF_joined}{DF_joined}
}
\description{
Checks if the NAs of the RAW columns are the same as the ones from the DIR columns
Important to catch errors when transforming data from RAW to DIR
}
