% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{check_project_and_results}
\alias{check_project_and_results}
\title{check_project_and_results
Check the project tasks and compare with the csv results to see if there are results or tasks missing}
\usage{
check_project_and_results(participants, folder_protocol, folder_results)
}
\arguments{
\item{participants}{.}

\item{folder_protocol}{.}

\item{folder_results}{.}
}
\description{
check_project_and_results
Check the project tasks and compare with the csv results to see if there are results or tasks missing
}
