% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{check_trialids}
\alias{check_trialids}
\title{check_trialids}
\usage{
check_trialids(local_folder_protocol, show_all_messages = FALSE)
}
\arguments{
\item{local_folder_protocol}{.}

\item{show_all_messages}{.}
}
\value{
cli messages with information
}
\description{
Checks that trialid's of an experiment in a folder follow the standard expected rules
}
