% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{clean_names_analysis}
\alias{clean_names_analysis}
\title{clean_names_analysis
Clean names of columns to use in output tables and plots}
\usage{
clean_names_analysis(DF)
}
\arguments{
\item{DF}{}
}
\value{
The same data frame after renaming the DIRd, DIRt and other similar variables
}
\description{
clean_names_analysis
Clean names of columns to use in output tables and plots
}
\examples{
clean_names_analysis(tibble::tibble(IdParticipant = 1, TASK_SimpleThing_DIRd = 1, TASK3_NotSoSimple_DIRt = 1))
}
