% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{cli_message}
\alias{cli_message}
\title{cli_message
Create standardize cli messages}
\usage{
cli_message(
  var_used = NULL,
  h1_title = NULL,
  h2_title = NULL,
  info = NULL,
  success = NULL,
  danger = NULL,
  details = NULL,
  list = NULL
)
}
\arguments{
\item{var_used}{if using a {variable}, need to include it here}

\item{h1_title}{title}

\item{h2_title}{subtitle}

\item{info}{info}

\item{success}{cli_success}

\item{danger}{cli_danger}

\item{details}{details}

\item{list}{list of elements}
}
\value{
A cli message
}
\description{
cli_message
Create standardize cli messages
}
\examples{
cli_message(h1_title = "title")
}
