% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_to_other_projects.R
\name{copy_to_other_projects}
\alias{copy_to_other_projects}
\title{copy_to_other_projects}
\usage{
copy_to_other_projects(
  file_name,
  overwrite_str = FALSE,
  check = TRUE,
  folder_to_search
)
}
\arguments{
\item{file_name}{.}

\item{overwrite_str}{.}

\item{check}{.}

\item{folder_to_search}{.}
}
\value{
copies file
}
\description{
Copy specific files to other projects
}
