% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{count_responses}
\alias{count_responses}
\title{count_responses}
\usage{
count_responses(DF, max_length_strings = 50)
}
\arguments{
\item{DF}{.}

\item{max_length_strings}{.}
}
\value{
Plots and tables
}
\description{
count_responses
}
