% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clean_data.R
\name{create_clean_data}
\alias{create_clean_data}
\title{create_clean_data}
\usage{
create_clean_data(DF_raw, save_output = TRUE, check_duplicates = TRUE)
}
\arguments{
\item{DF_raw}{.}

\item{save_output}{.}

\item{check_duplicates}{.}
}
\description{
create_clean_data
}
