% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{create_codebook}
\alias{create_codebook}
\title{create_codebook
Extracts information from the prepare_TASK.R files to create a codebook. Used in report_DF_clean.Rmd}
\usage{
create_codebook(task)
}
\arguments{
\item{task}{A prepare_TASK.R file}
}
\value{
A DF with correction information of a prepare_TASK() function
}
\description{
create_codebook
Extracts information from the prepare_TASK.R files to create a codebook. Used in report_DF_clean.Rmd
}
\examples{
create_codebook(task = system.file("R_tasks", "prepare_AIM.R", package = "jsPsychHelpeR"))
}
