% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DF_analysis.R
\name{create_DF_analysis}
\alias{create_DF_analysis}
\title{create_DF_analysis}
\usage{
create_DF_analysis(DF_joined, last_task, save_output = TRUE, DVars = c(""))
}
\arguments{
\item{DF_joined}{DF}

\item{last_task}{name of last task}

\item{save_output}{TRUE / FALSE}

\item{DVars}{vector with variable names}
}
\value{
DF_analysis
}
\description{
create_DF_analysis
}
