% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diccionary_tasks.R
\name{create_diccionary_tasks}
\alias{create_diccionary_tasks}
\title{create_diccionary_tasks}
\usage{
create_diccionary_tasks(DF_clean)
}
\arguments{
\item{DF_clean}{DF_clean}
}
\value{
Writes outputs/data/DICCIONARY_tasks.csv
}
\description{
Create DICCIONARY with long and short names for the tasks in the protocol
}
