% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dictionary_tasks.R
\name{create_dictionary_tasks}
\alias{create_dictionary_tasks}
\title{create_dictionary_tasks}
\usage{
create_dictionary_tasks(DF_clean)
}
\arguments{
\item{DF_clean}{DF_clean}
}
\value{
Writes outputs/data/DICTIONARY_tasks.csv
}
\description{
Create DICTIONARY with long and short names for the tasks in the protocol
}
