% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{create_formulas}
\alias{create_formulas}
\title{create_formulas
It creates the standardized code necessary for the dimensions or total scores calculation in the prepare_TASK scripts
PROBABLY NOT NEEDED, as we SHOULD do all this automatically in get_dimensions_google_doc()
Maybe useful when there are some non-standard elements to the calculations? e.g. See prepare_CS}
\usage{
create_formulas(type, functions = "sum", dimensions = NULL)
}
\arguments{
\item{type}{One of: c("dimensions_DIR", "dimensions_STD", "dimensions_REL", "total_DIR", "total_REL", "total_STD")}

\item{functions}{One of: c("sum"|"mean")}

\item{dimensions}{character vector with the dimensions names}
}
\value{
prints the formula for the prepare_TASK() function
}
\description{
create_formulas
It creates the standardized code necessary for the dimensions or total scores calculation in the prepare_TASK scripts
PROBABLY NOT NEEDED, as we SHOULD do all this automatically in get_dimensions_google_doc()
Maybe useful when there are some non-standard elements to the calculations? e.g. See prepare_CS
}
\examples{
 create_formulas(type = "dimensions_DIR", functions = "sum", dimensions = c("ONE"))
}
