% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{create_raw_long}
\alias{create_raw_long}
\title{create_raw_long}
\usage{
create_raw_long(
  DF_clean,
  short_name_scale,
  numeric_responses = FALSE,
  is_experiment = FALSE,
  keep_time = FALSE,
  help_prepare = FALSE
)
}
\arguments{
\item{DF_clean}{DF_clean in a prepare_TASK}

\item{short_name_scale}{Short name of scale}

\item{numeric_responses}{TRUE / FALSE}

\item{is_experiment}{TRUE / FALSE}

\item{keep_time}{TRUE / FALSE keep datetime in raw_long}

\item{help_prepare}{TRUE / FALSE}
}
\value{
A long and clean version of DF_clean for a specific task
}
\description{
Create long DF for a specific task (short_name_scale). Also cleans the DF and if it's an experiment, makes sure names are OK.
}
