% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{create_targets_file}
\alias{create_targets_file}
\title{Create _targets.R file from a protocol folder}
\usage{
create_targets_file(pid = 0, folder, dont_ask = FALSE)
}
\arguments{
\item{pid}{project id}

\item{folder}{protocol folder}

\item{dont_ask}{do everything without asking for user input}
}
\value{
Creates a _targets.R file
}
\description{
Create _targets.R file from a protocol folder
}
\examples{
\dontrun{
create_targets_file(pid = 999, 
                    folder = "/home/emrys/Downloads/jsPsychHelpeRtest/",
                    dont_ask = FALSE)
}
}
