% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_duplicates.R
\name{delete_duplicates}
\alias{delete_duplicates}
\title{delete_duplicates}
\usage{
delete_duplicates(folder, check = TRUE, keep_which = "older")
}
\arguments{
\item{folder}{.}

\item{check}{TRUE/FALSE}

\item{keep_which}{"older"/"newer"}
}
\description{
Checks or deletes duplicate files, keeping the older file.
}
