% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{move_sensitive_tasks_to_vault}
\alias{move_sensitive_tasks_to_vault}
\title{move_sensitive_tasks_to_vault
Move sensitive files to .vault/data_vault}
\usage{
move_sensitive_tasks_to_vault(pid, folder, sensitive_tasks = c(""))
}
\arguments{
\item{pid}{.}

\item{folder}{.}

\item{sensitive_tasks}{.}
}
\description{
move_sensitive_tasks_to_vault
Move sensitive files to .vault/data_vault
}
