% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{number_items_tasks}
\alias{number_items_tasks}
\title{number_items_tasks
Creates a tibble with information about how many RAW items each task in DF_joined has}
\usage{
number_items_tasks(DF_joined)
}
\arguments{
\item{DF_joined}{.}
}
\description{
number_items_tasks
Creates a tibble with information about how many RAW items each task in DF_joined has
}
