% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{prepare_helper}
\alias{prepare_helper}
\title{helper to correct tasks}
\usage{
prepare_helper(DF_long_RAW, show_trialid_questiontext = FALSE)
}
\arguments{
\item{DF_long_RAW}{dataframe}

\item{show_trialid_questiontext}{TRUE/FALSE}
}
\description{
helper to correct tasks
}
