% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read_data()}
\usage{
read_data(input_files, anonymize = FALSE, save_output = TRUE, workers = 1)
}
\arguments{
\item{input_files}{files to read}

\item{anonymize}{TODO}

\item{save_output}{should save output?}

\item{workers}{number of threads for data.table::fread}
}
\description{
Read raw data and prepare a global DF
}
\author{
gorkang
}
