% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{read_zips}
\alias{read_zips}
\title{read_zips
Function to unzip and read csv files}
\usage{
read_zips(
  input_files,
  workers = 1,
  unzip_dir = file.path(dirname(input_files), sprintf("csvtemp_\%s", sub(".zip", "",
    basename(input_files)))),
  silent = FALSE,
  do_cleanup = TRUE,
  only_list = FALSE
)
}
\arguments{
\item{input_files}{.}

\item{workers}{.}

\item{unzip_dir}{.}

\item{silent}{.}

\item{do_cleanup}{.}

\item{only_list}{.}
}
\description{
read_zips
Function to unzip and read csv files
}
