% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{save_files}
\alias{save_files}
\title{create_raw_wide}
\usage{
save_files(
  DF,
  short_name_scale,
  is_scale = TRUE,
  is_sensitive = FALSE,
  output_formats = c("csv", "csv2")
)
}
\arguments{
\item{DF}{dataframe}

\item{short_name_scale}{short name of scale}

\item{is_scale}{= TRUE}

\item{is_sensitive}{= TRUE}

\item{output_formats}{One or more of the following: csv, csv2 or rds}
}
\description{
Save files
}
\details{
Save a DF in one or more output formats to output/data or .vault/outputs/data/
}
