% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{set_permissions_google_drive}
\alias{set_permissions_google_drive}
\title{set_permissions_google_drive}
\usage{
set_permissions_google_drive(pid, email_IP)
}
\arguments{
\item{pid}{pid of project}

\item{email_IP}{email of Principal Researcher to give reading permissions to data in google drive}
}
\description{
Changes permission in a google drive folder
}
