% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{setup_folders}
\alias{setup_folders}
\title{setup_folders
Extract jsPsychHelpeR.zip and make sure we have all necessary folders in a specific location}
\usage{
setup_folders(pid, folder, extract_zip = FALSE)
}
\arguments{
\item{pid}{project id}

\item{folder}{destination folder for the project}

\item{extract_zip}{If TRUE, extracts jsPsychHelpeR.zip to folder}
}
\description{
setup_folders
Extract jsPsychHelpeR.zip and make sure we have all necessary folders in a specific location
}
