% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{show_progress_pid}
\alias{show_progress_pid}
\title{show_progress_pid}
\usage{
show_progress_pid(
  pid = 3,
  files_vector,
  last_task = "Goodbye",
  goal = 100,
  DEBUG = FALSE
)
}
\arguments{
\item{pid}{project id}

\item{files_vector}{vector of files}

\item{last_task}{short name of last task}

\item{goal}{number of participants}

\item{DEBUG}{TRUE / FALSE}
}
\value{
Table, plot, DF... about the progress of data collection
}
\description{
Show progress of data collection of a project
}
