% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{snapshot_check}
\alias{snapshot_check}
\title{snapshot_check
Check all changing snapshots. For some reason testthat::snapshot_review() does
not seem to work when multiple images change. With snapshot_check, we use
testthat:::review_app to check all .new snapshots}
\usage{
snapshot_check(snapshot_location = "testthat/_snaps/snapshots/")
}
\arguments{
\item{snapshot_location}{Location of the snapshots, by default "testthat/_snaps/snapshots/"}
}
\value{
Open a shiny app
}
\description{
snapshot_check
Check all changing snapshots. For some reason testthat::snapshot_review() does
not seem to work when multiple images change. With snapshot_check, we use
testthat:::review_app to check all .new snapshots
}
