% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{standardized_names}
\alias{standardized_names}
\title{standardized_names}
\usage{
standardized_names(short_name_scale, dimensions = "", help_names = FALSE)
}
\arguments{
\item{short_name_scale}{short name of scale}

\item{dimensions}{c("DIMENSION1", "DIMENSION2"): list of names for the dimensions}

\item{help_names}{TRUE prints a message with the instructions}
}
\value{
Creates names_list, which can be used in the prepare_TASK scripts to
create columns with standardized names for direct scores, total scores, etc.
}
\description{
Standardized names for Direct scores, dimensions and scales
}
