% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_server_local.R
\name{sync_server_local}
\alias{sync_server_local}
\title{sync_server_local
Standalone function to Download/Upload files from server to local folder or vice-versa}
\usage{
sync_server_local(
  server_folder,
  local_folder,
  direction,
  only_test = TRUE,
  exclude_csv = FALSE,
  delete_nonexistent = FALSE,
  ignore_existing = FALSE,
  dont_ask = FALSE,
  all_messages = TRUE,
  credentials_file = NULL
)
}
\arguments{
\item{server_folder}{server folder}

\item{local_folder}{local folder}

\item{direction}{should be either 'server_to_local' or 'local_to_server'}

\item{only_test}{TRUE/FALSE}

\item{exclude_csv}{DO NOT INCLUDE DATA}

\item{delete_nonexistent}{Delete files locally if they are NOT in server anymore}

\item{ignore_existing}{If TRUE, does not overwrite existing files even if they are newer. Good for .data/, Bad for rest}

\item{dont_ask}{For the initial setup, do not ask and proceed! DANGER}

\item{all_messages}{show all messages}

\item{credentials_file}{Path to .credentials file. Usually: \code{.vault/.credentials}}
}
\value{
Sync folders
}
\description{
sync_server_local
Standalone function to Download/Upload files from server to local folder or vice-versa
}
