% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{update_data}
\alias{update_data}
\title{update_data
Download 'pid/.data/' server folder to 'data/pid' using rsync. If there are sensitive tasks, move those to .vault/data_vault}
\usage{
update_data(pid, folder)
}
\arguments{
\item{pid}{protocol id}

\item{folder}{project folder. Data will be downloaded to folder/data/pid}
}
\value{
Downloads data from server to the folder
}
\description{
update_data
Download 'pid/.data/' server folder to 'data/pid' using rsync. If there are sensitive tasks, move those to .vault/data_vault
}
