% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{zip_files}
\alias{zip_files}
\title{zip_files
zip files of a folder}
\usage{
zip_files(folder_files, zip_name, remove_files = FALSE)
}
\arguments{
\item{folder_files}{.}

\item{zip_name}{.}

\item{remove_files}{.}
}
\description{
zip_files
zip files of a folder
}
