% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{auto_reliability}
\alias{auto_reliability}
\title{auto_reliability}
\usage{
auto_reliability(
  DF,
  short_name_scale = short_name_scale_str,
  items = NULL,
  min_rdrop = 0.2
)
}
\arguments{
\item{DF}{dataframe}

\item{short_name_scale}{short name of scale}

\item{items}{items to check}

\item{min_rdrop}{rdrop threshold}
}
\value{
Saves a DF and outputs information about the item selection
}
\description{
Checks Cronbach's alpha and select variables based on a min r.drop criteria
}
