% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{check_NAs}
\alias{check_NAs}
\title{check_NAs}
\usage{
check_NAs(DF_joined)
}
\arguments{
\item{DF_joined}{.}
}
\description{
Checks if the NAs of the RAW calculation are the same as the ones from the PROC calculation
}
\details{
Important to catch errors when transforming data from RAW to PROCESSED
}
\author{
gorkang
}
