% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{clean_renv_cache}
\alias{clean_renv_cache}
\title{clean_renv_cache
Gets rid of cache packages not in lib}
\usage{
clean_renv_cache(
  lib_location = "renv/lib/R-4.3/x86_64-pc-linux-gnu/",
  cache_location = "renv/cache/v5/R-4.3/x86_64-pc-linux-gnu"
)
}
\value{
Cleans the renv cache
}
\description{
clean_renv_cache
Gets rid of cache packages not in lib
}
