% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{create_codebook}
\alias{create_codebook}
\title{create_codebook
Extracts information from the prepare_TASK.R files to create a codebook}
\usage{
create_codebook(tasks, number)
}
\arguments{
\item{tasks}{.}

\item{number}{.}
}
\description{
create_codebook
Extracts information from the prepare_TASK.R files to create a codebook
}
