% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{create_new_task}
\alias{create_new_task}
\title{create_new_task
Create a new prepare_TASK.R file from prepare_TEMPLATE.R replacing TEMPLATE by the short name of the new task}
\usage{
create_new_task(
  short_name_task,
  overwrite = FALSE,
  get_info_googledoc = FALSE,
  destination = "R_tasks"
)
}
\arguments{
\item{short_name_task}{short name of the task}

\item{overwrite}{FALSE / TRUE}

\item{get_info_googledoc}{FALSE / TRUE}

\item{destination}{"R_tasks"}
}
\value{
Creates a R_tasks/prepare_TASK.R using the prepare_TEMPLATE.R
}
\description{
create_new_task
Create a new prepare_TASK.R file from prepare_TEMPLATE.R replacing TEMPLATE by the short name of the new task
}
