% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{create_number_series}
\alias{create_number_series}
\title{create_number_series
Create a number series from a vector of numbers including individual numbers
and intervals such as c(1, "3-8", 11, "22-24")}
\usage{
create_number_series(numbers_RAW)
}
\arguments{
\item{numbers_RAW}{.}
}
\description{
create_number_series
Create a number series from a vector of numbers including individual numbers
and intervals such as c(1, "3-8", 11, "22-24")
}
