% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{create_raw_long}
\alias{create_raw_long}
\title{create_raw_long}
\usage{
create_raw_long(
  DF_clean,
  short_name_scale,
  numeric_responses = FALSE,
  is_experiment = FALSE,
  help_prepare = FALSE
)
}
\arguments{
\item{DF_clean}{.}

\item{short_name_scale}{.}

\item{numeric_responses}{.}

\item{is_experiment}{.}

\item{help_prepare}{.}
}
\description{
Create long DF for a specific task (short_name_scale)
}
