% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_duplicates.R
\name{delete_duplicates}
\alias{delete_duplicates}
\title{delete_duplicates}
\usage{
delete_duplicates(folder, check = TRUE, keep_which = "older")
}
\arguments{
\item{folder}{A folder with a zip or csv files from a jsPsychMaker protocol}

\item{check}{If TRUE, only checks, but does not make changes}

\item{keep_which}{"older"/"newer"}
}
\value{
A list with information
}
\description{
Checks or deletes duplicate files, keeping the older file.
}
\examples{
delete_duplicates(folder = system.file("extdata/",package = "jsPsychHelpeR"))
}
