% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dimensions_googledoc.R
\name{get_dimensions_googledoc}
\alias{get_dimensions_googledoc}
\title{get_dimensions_googledoc
Get information from the main jsPsychR GoogleSheet to help creating a new task}
\usage{
get_dimensions_googledoc(
  short_name_text,
  google_username = "gorkang@gmail.com",
  google_sheet = "NEW"
)
}
\arguments{
\item{short_name_text}{.}

\item{google_username}{.}

\item{google_sheet}{.}
}
\description{
get_dimensions_googledoc
Get information from the main jsPsychR GoogleSheet to help creating a new task
}
