% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{get_zip}
\alias{get_zip}
\title{Get and zip the data, or a full jsPsychMakeR protocol without the data, to keep it as a backup}
\usage{
get_zip(
  pid,
  what,
  where = NULL,
  list_credentials = NULL,
  dont_ask = TRUE,
  all_messages = FALSE,
  tempdir_location = NULL
)
}
\arguments{
\item{pid}{project id}

\item{what}{Should be one of c("data", "protocol")}

\item{where}{Where to leave the zip file with the data. Leave empty to save to \verb{SHARED-data/pid/}}

\item{list_credentials}{list with the credentials. Usually source(".vault/.credentials")}

\item{dont_ask}{TRUE / FALSE}

\item{all_messages}{Show all rsync messages? TRUE / FALSE}

\item{tempdir_location}{You can choose a tempdir_location (for example, to extract contents of an existing zip and sync only new files)}
}
\value{
A zip file
}
\description{
We use this internal function to save the data of the Running protocols
(according to the Google sheet "Codebook protocolos") to a folder SHARED with
the project's PI's
}
