% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{get_zip}
\alias{get_zip}
\title{get_zip
Get and zip a the data or a full jsPsychMakeR protocol without the data to keep it as a backup}
\usage{
get_zip(pid, what, where = NULL, list_credentials = NULL, dont_ask = TRUE)
}
\arguments{
\item{pid}{project id}

\item{what}{data/protocol}

\item{where}{.}

\item{list_credentials}{list with the credentials. Usually source(".vault/.credentials")}

\item{dont_ask}{TRUE / FALSE}
}
\value{
A zip file
}
\description{
get_zip
Get and zip a the data or a full jsPsychMakeR protocol without the data to keep it as a backup
}
