% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_input_files.R
\name{list_input_files}
\alias{list_input_files}
\title{list_input_files
Lists all input csv files or single zip file in a data folder.}
\usage{
list_input_files(pid_target)
}
\arguments{
\item{pid_target}{.}
}
\description{
list_input_files
Lists all input csv files or single zip file in a data folder.
}
