% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{parse_filename}
\alias{parse_filename}
\title{parse_filename}
\usage{
parse_filename(DF, separator = "_")
}
\arguments{
\item{DF}{DF with a list of}

\item{separator}{.}
}
\description{
Parse input files to get the essential pieces encoded in the filename column
}
