% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{prepare_helper}
\alias{prepare_helper}
\title{prepare_helper}
\usage{
prepare_helper(DF_long_RAW, show_trialid_questiontext = FALSE)
}
\arguments{
\item{DF_long_RAW}{DF_long_RAW created by create_raw_long()}

\item{show_trialid_questiontext}{TRUE / FALSE}
}
\value{
A number of tables, DT, etc.
}
\description{
Helper to correct tasks. Shows trialid's, responses, etc. to help correcting a task
}
