% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{read_csv_or_zip}
\alias{read_csv_or_zip}
\title{read_csv_or_zip
Read input files, adapting to multiple csv's or a single zip}
\usage{
read_csv_or_zip(input_files, workers = 1, only_list = FALSE)
}
\arguments{
\item{input_files}{multiple csv's or a single zip}

\item{workers}{how many cores to use}

\item{only_list}{.}
}
\description{
read_csv_or_zip
Read input files, adapting to multiple csv's or a single zip
}
