% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(
  input_files,
  is_sensitive = FALSE,
  save_output = TRUE,
  workers = 1,
  location_duplicates = NULL
)
}
\arguments{
\item{input_files}{A character vector of files to read}

\item{is_sensitive}{If the data contains sensitive information, select TRUE to save in .vault. TRUE / FALSE}

\item{save_output}{TRUE / FALSE}

\item{workers}{number of threads for data.table::fread}

\item{location_duplicates}{If there are duplicate trialid's, a csv with details will be saved here}
}
\value{
A DF
}
\description{
Read raw data and prepare a global DF
}
