% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_minimal.R
\name{read_zips}
\alias{read_zips}
\title{read_zips
Function to unzip and read csv files. This is used in read_csv_or_zip()}
\usage{
read_zips(
  input_files,
  workers = 1,
  unzip_dir = file.path(tempdir(), sprintf("csvtemp_\%s", sub(".zip", "",
    basename(input_files)))),
  silent = FALSE,
  do_cleanup = TRUE,
  only_list = FALSE
)
}
\arguments{
\item{input_files}{A zip, tar or xz file}

\item{workers}{Workers for data.table::fread}

\item{unzip_dir}{Where to unzip (if do_cleanup == TRUE, the folder will be deleted)}

\item{silent}{Show feedback FALSE / TRUE}

\item{do_cleanup}{Delete temp folder TRUE / FALSE}

\item{only_list}{Only list the files, but do not read them}
}
\value{
Either a list of files or a dataframe after reading the files
}
\description{
read_zips
Function to unzip and read csv files. This is used in read_csv_or_zip()
}
\examples{
read_zips(system.file("extdata/", "999.zip", package = "jsPsychHelpeR"))
}
