% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_initial_setup.R
\name{run_initial_setup}
\alias{run_initial_setup}
\title{run_initial_setup
Run initial setup for a specific project}
\usage{
run_initial_setup(
  pid,
  download_files = FALSE,
  data_location = NULL,
  download_task_script = FALSE,
  folder = "~/Downloads/jsPsychHelpeRtest",
  sensitive_tasks = c(""),
  dont_ask = FALSE
)
}
\arguments{
\item{pid}{project id}

\item{download_files}{should download the data files? (requires server credentials) TRUE/FALSE}

\item{data_location}{local folder where the raw data for the project is}

\item{download_task_script}{should download the task scripts? (requires server credentials) TRUE/FALSE}

\item{folder}{location for the project}

\item{sensitive_tasks}{short names of the sensitive tasks in the protocol, if any}

\item{dont_ask}{answer YES to all questions}
}
\value{
Opens a new RStudio project
}
\description{
run_initial_setup
Run initial setup for a specific project
}
