% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_extra.R
\name{zip_files}
\alias{zip_files}
\title{zip_files
zip files of a folder}
\usage{
zip_files(folder_files, zip_name, remove_files = FALSE)
}
\arguments{
\item{folder_files}{folder to zip}

\item{zip_name}{name of output zip}

\item{remove_files}{remove source files TRUE / FALSE}
}
\description{
zip_files
zip files of a folder
}
